+++ process/doc.go
@@ -1,0 +1,2 @@
+// Process contains various utils process handling
+package process